Rename lib.yang.schema2 to lib.yang.schema
* src/lib/yang/schema.lua: Rename from schema2.lua.
* src/lib/yang/yang.lua: Just re-export things from schema and data
  modules.  Document via README.md.
